changed mu back to float instead of double...

+++ wavenet/ops.py
@@ -64,7 +64,7 @@ def causal_conv(value, filter_, dilation, name='causal_conv'):
 def mu_law_encode(audio, quantization_channels):
     """Quantizes waveform amplitudes."""
     with tf.name_scope('encode'):
-        mu = tf.to_double(quantization_channels - 1)
+        mu = tf.to_float(quantization_channels - 1)
         # Perform mu-law companding transformation (ITU-T, 1988).
         # Minimum operation is here to deal with rare large amplitudes caused
         # by resampling.
